[heap] Reserve a word at the end of each memory chunk
Currently we allocate an additional cell for marking bitmap to account
for a one-word-filler object at the very end of a memory chunk.
This makes the bitmap size 8KB + 4 bytes, which does not work well
with some malloc implementations and leads to large memory usages.

This CL implements an alternative solution by disallowing allocation
of the last word in each memory chunk. It is done by setting area_end
accordingly. The CL also fixes the existing code that assumed that
area_end is page/cell-aligned.

Change-Id: I55d9196701f41e528237a13464c233248dae7e81