[TurboFan] Print more information in logs related to dependency change
When marking the function for a deopt due a dependency change, we
only print the information about the code object. This isn't very
useful when looking at the output of --trace-deopt. This cl also adds
SharedFunctionInfo which makes it easier to see which function got
deoptimized. Also adds these events to the log file so they can be used
in profview to print more accurate deoptimization information.

Change-Id: I1b7a88b5cc2a1ad1ebb0863cf8a4f7d2e151b76a