[liftoff] Fix incoming register pairs
For register pairs we have the constraint that the upper half has a
higher register code than the lower half. This is enforced in
{GetUnusedRegister} since https://crrev.com/c/2168875, and is relied on
in the LiftoffAssembler on arm.
For incoming parameter though, we just took the two parameter registers
and used them as a register pair. This violated the constraint, leading
to code generation issues later.

This CL fixes this by
1) checking for correct register pair ordering in
   LiftoffRegister::ForPair, and
2) swapping incoming registers if their order is inverted.

R=thibaudm@chromium.org

Bug: chromium:1101304
Change-Id: I4bf6ec578dc571ad0483619aa5a464c612f0f6c6