[TurboProp] Add PendingOperand for use in fast register allocater.
Adds a PendingOperand operand type. This type contains a pointer to another
PendingOperand, enabling it to be used as a threaded list. This is used by
the fast register allocator to chain together lists of operands which haven't
yet been allocated with a specific allocated operand, but will all be replaced
with the same operand once the decision is made.

BUG=v8:9684

Change-Id: I5d8150f3f26549a747a2e89e32e31135e89dff9c