[TurboProp] Add initial DefineOutput phase for fast register allocator
Adds the first phase of the fast register allocator, which runs through

the instruction stream and defines a VirtualRegisterData for each virtual
register based on how that virtual register is produced. Also adds logic
to pipeline.cc to allocate and use a FastRegistorAllocatorData for use
throughout the fast register allocation phases.

BUG=v8:9684

Change-Id: I2f4533467346d5f3fdf50a0a1fedd7e4082f0187