[TurboProp] Add reference map population to fast reg alloc.
Adds support for poplulating reference maps to the fast
register allocator. In order to calculate whether a stack slot
is live at a given instruction, we use the dominator tree to
build a bitmap of blocks which are dominated by each block.
When a variables spill operand is live for any blocks dominated
by the block it was defined in, until that spill operand's last use.

BUG=v8:9684

Change-Id: I772374599ef916f57d82d468f66429e32c712ddf