[WIP][sandbox][x64] Access external pointer through a table
This change moves external pointers into a separate table datastructure
and turns external pointers in heap objects into indices into that
table.

This CL implements one of two possible ownership models for the table
entries. With this one, every heap objects owns its table entries. The
entries are allocated when the owning object is allocated and they are
freed when the object is garbage collected.

As such, setting external pointer fields does not require allocation of
table entries and the memory management of the table is simpler. On the
other hand, table indices cannot be shared between different objects and
table entries are allocated even when the external pointers are never
set (although for the common cases it should be possible to only
allocate table entries when they are needed).

This CL does not yet implement freeing of external pointer table entires
(this has to happen in object finalizers) and it also does not correctly
handle EmbedderDataSlots.

Bug: v8:10391
Change-Id: I4d37785295c25a7d1dcbc9871dd5887b9d788a4f