[wasm] Don't automatically remove breakpoints
Out of the four listed reasons, two are obsolete now, since stepping
does not modify the code table any more. The first reason also cannot
happen any more, since removing a breakpoint will also patch the stack
of the isolate, and even if not, removing a breakpoint which is not set
will have no effect (not even a stack rewriting). The same applies for
the last listed reason.

So overall, this code is not needed any more, since it has no effect.

R=thibaudm@chromium.org

Bug: v8:10611
Change-Id: I773ac2515b66e8bf66cb035c315db552024485c3