[wasm] Remove breakpoint info when tiering up
If we keep the breakpoint information, a second isolate that later sets
the same breakpoint will see that the breakpoint already exists, and
will not set it again, even though the code containing the breakpoint
has been replaced at that point.
This fixes a flake in the debug/wasm/breakpoints test.

R=thibaudm@chromium.org

Bug: v8:10611
Change-Id: I3ae5f7da06c5850650cad9b3965d69edad9cf502