[regalloc] Avoid a spill & reload in some loops
Current behavior:

Live ranges are collected into bundles, where each bundle represents a
set of values that can share the same stack slot. V8 attempts to bundle
phi values together with their predecessors, but doing so is not always
possible. In cases where a loop-top phi value can't be bundled together
with all of its predecessors defined in the loop, the corresponding
TopLevelLiveRange is marked as SpillAtLoopHeaderNotBeneficial. If
SpillAtLoopHeaderNotBeneficial is set, then FindOptimalSpillingPos will
not choose to mark the range as spilled starting at its definition, and
will instead split the range at some later point to introduce a spilled
chunk. (If the range is marked as spilled starting at its definition,
then that means the value must be stored in the correct stack slot
by each predecessor block of the loop top.) The goal of this mechanism
is to avoid stack-to-stack moves at the ends of the loop.

However, sometimes the phi value gets a spill inserted at the top of
the loop anyway, which means we didn't gain much by avoiding the
stack-to-stack move, and we may have unnecessarily harmed the
performance of predecessors which were successfully bundled with the
loop-top phi value.

This change:

In this change, I propose that when checking whether marking the range
as spilled from its start might be beneficial, we should take into
account what the alternative is. If the alternative split position is
near the loop-top block and along the only non-deferred execution path,
then marking the value as spilled from its start is still likely better
than introducing a split at the alternative position, regardless of what
the SpillAtLoopHeaderNotBeneficial flag says.

This improves the performance of the Mono interpreter by about 2%.

Bug: v8:10606
Change-Id: Id0c7720896d27731f67976d3f9aed74418164835