[regalloc] Avoid an unnecessary load & store in some loops
Current behavior:

Live ranges are collected into bundles, where each bundle represents a
set of values that can share the same stack slot. V8 attempts to bundle
phi values together with their predecessors, but doing so is not always
possible because the values' lifetimes might overlap. In cases where a
loop-top phi value can't be bundled together with all of its back-edge
input values, the corresponding TopLevelLiveRange is marked as
SpillAtLoopHeaderNotBeneficial. If SpillAtLoopHeaderNotBeneficial is
set, then FindOptimalSpillingPos will not choose to mark the range as
spilled starting at its definition, and will instead split the range at
some later point to introduce a spilled chunk. (If the range is marked
as spilled starting at its definition, that means the value must be
stored in the correct stack slot by each predecessor block of the loop
top.) The goal of this mechanism is to avoid stack-to-stack moves at the
ends of the loop.

However, this heuristic can be harmful when there are multiple back-edge
inputs to the loop-top phi value, and only some of them can't be
bundled. For inputs that were sucessfully bundled, we can end up with a
load at the end of the loop followed by a store at the beginning right
back to the same stack slot.

This change:

I propose we change the heuristic so that SpillAtLoopHeaderNotBeneficial
is set only if none of the back-edge inputs can be bundled together with
the phi value. This improves the Mono interpreter by about 6%, and I
checked that this doesn't regress the ArrayLiteralInitialSpread* tests
on ia32 which SpillAtLoopHeaderNotBeneficial was introduced to fix.

Bug: v8:10606
Change-Id: Ie23a3d7a731a6cfb1f95acfccbf079bbe30a40b8