[regalloc] Loop-top values can be backedges too
When counting which backedges conflict with a loop-top phi value, we
should include values introduced at the start of the loop.

I don't expect this change to make performance differences on its own,
but it is a step toward changing the heuristic so that we're less likely
to introduce unnecessary load/store pairs across the backedge.

Bug: v8:10606
Change-Id: I299e388b0b964573119ba0b775d50f398c467c46