[wasm] Scan stack in foreground code GC
When a WasmGCForegroundTask starts during runMessageLoopOnPause, the
stack may contain Wasm code. In this case we need to scan the stack to
report live code.

This is already caught by some tests when RemoveBreakpoint is
implemented correctly, which will be done in a separate CL, so no
regression test is added here.

R=clemensb@chromium.org

Bug: v8:10337
Change-Id: I308145b9c2cf69f0100853be47ec9e30ce79a876