[turboprop] Infer elements kind from handler for dynamic map checks
JSCallReducer uses information from the earlier map checks to gather
more information about calls to array builtins to handle them better.

For example when we have
a.push(10);
Interpreter generates a named load (for a.push) followed by a call. At
the call site we don't have information about 'a'. JSCallReducer walks
up the effect chain (in MapInference) and tries to find information
about 'a' from earlier CheckMaps operators. one of the important
information it gathers is about the elements kind of the array and if
the array allows fast resizes. With this information it is able to
inline these builtins for better performance.

When using dynamic map checks for named property loads such information
isn't available since dynamic map checks don't include maps and it isn't
safe to infer any information from maps. This causes performance
regressions in some of the pages that rely on fast array accesses.

To address these performance issues, dynamic map checks uses the
information from that is encoded in the handlers. An earlier cl:
https://chromium-review.googlesource.com/c/v8/v8/+/2247652
encoded the elements kind into unused bits of IC handlers. This cl
1. encodes the elements kind only if the JSArray supports fast array
   resize
2. Uses the information encoded in the handlers to provide
   information about JSArray's elements kinds in JSCallReducer so it
   could inline array builtins when possible. It is safe to use the
   information from the handlers, since dynamic map checks also check
   for the right handler.

Bug: v8:10582
Change-Id: I653de3f5e49d775ec7cce19215073074322abf65