[wasm-simd][scalar-lowering] Fix all_true lowering
AllTrue requires the input to be of the respective type, but the
final result is always a Int32x4 with a single node. So keep
the replacement type of AllTrue (and AnyTrue) nodes to be Int32x4,
and use a helper method to decide what SimdType the input should
be replaced with.

Also split up AnyTrue and AllTrue for readability.

Bug: v8:10507
Change-Id: I58ca50ffffcbca3ec77bbae1371ddd179925fc96